srt_whisper_lite/requirements_api.txt
# SRT GO API 服務依賴
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6

# 現有核心依賴
faster-whisper>=0.10.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
noisereduce>=3.0.0

# 可選依賴
Pillow>=10.0.0
requests>=2.31.0